moved some values

+++ Assets/Scripts/Player/Resources.cs
@@ -7,14 +7,14 @@ public class Resources : MonoBehaviour
 {
 	[SerializeField] TextMeshProUGUI label;
 
-	int coins = 0;
+	double coins = 0;
 
 	void Start()
 	{
 		label.text = "0";
 	}
 
-	public void AddCoins(int coins)
+	public void AddCoins(double coins)
 	{
 		this.coins += coins;
 		label.text = this.coins.ToString();
